Return int64 from count convenience methods (#31)
* Return int64 from GetListenerCount, GetPlaycount, GetUserPlaycount

These methods returned float64 despite wrapping int64 fields from the
Info structs. Listener and play counts are always integers; float64
silently loses precision above 2^53. Now they return int64 directly.


* Document breaking changes in CHANGELOG

Add entries for the int64 return type change and WSError.Status type
change, along with the performance and correctness fixes from this
release cycle.


---------

+++ lastfm/artist.go
@@ -102,30 +102,30 @@ func (a *Artist) GetMBID(ctx context.Context) (string, error) {
 }
 
 // GetListenerCount returns the number of Last.fm listeners.
-func (a *Artist) GetListenerCount(ctx context.Context) (float64, error) {
+func (a *Artist) GetListenerCount(ctx context.Context) (int64, error) {
 	info, err := a.GetInfo(ctx)
 	if err != nil {
 		return 0, fmt.Errorf("Artist.GetListenerCount: %w", err)
 	}
-	return float64(info.Listeners), nil
+	return info.Listeners, nil
 }
 
 // GetPlaycount returns the total play count on Last.fm.
-func (a *Artist) GetPlaycount(ctx context.Context) (float64, error) {
+func (a *Artist) GetPlaycount(ctx context.Context) (int64, error) {
 	info, err := a.GetInfo(ctx)
 	if err != nil {
 		return 0, fmt.Errorf("Artist.GetPlaycount: %w", err)
 	}
-	return float64(info.Playcount), nil
+	return info.Playcount, nil
 }
 
 // GetUserPlaycount returns the play count for the authenticated user.
-func (a *Artist) GetUserPlaycount(ctx context.Context) (float64, error) {
+func (a *Artist) GetUserPlaycount(ctx context.Context) (int64, error) {
 	info, err := a.GetInfo(ctx)
 	if err != nil {
 		return 0, fmt.Errorf("Artist.GetUserPlaycount: %w", err)
 	}
-	return float64(info.UserPlaycount), nil
+	return info.UserPlaycount, nil
 }
 
 // GetSimilar returns artists similar to this one.
